# documention/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/national/organisation/query": {
      "post": {
        "operationId": "CompanyController_query",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/byType": {
      "post": {
        "operationId": "CompanyController_byType",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ByTypeDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/getOrganizations": {
      "post": {
        "operationId": "CompanyController_getOrganizations",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetOrganizationsRequest"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/organisation/queryNames": {
      "post": {
        "operationId": "CompanyController_queryNames",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/download": {
      "post": {
        "operationId": "CompanyController_getDownload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DataExportQueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/suspend": {
      "put": {
        "operationId": "CompanyController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationSuspendDto"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/activate": {
      "put": {
        "operationId": "CompanyController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationSuspendDto"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/approve": {
      "put": {
        "operationId": "CompanyController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationSuspendDto"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/reject": {
      "put": {
        "operationId": "CompanyController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrganisationRejectDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/findByIds": {
      "post": {
        "operationId": "CompanyController_findByCompanyId",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindOrganisationQueryDto"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/profile": {
      "get": {
        "operationId": "CompanyController_getCompany",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/organisation/update": {
      "put": {
        "operationId": "CompanyController_updateCompany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrganisationUpdateDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/countries": {
      "post": {
        "operationId": "CompanyController_getCountries",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "CompanyController_getAvailableCountries",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/organisation/regions": {
      "post": {
        "operationId": "CompanyController_getRegionList",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/organisation/addInvestment": {
      "post": {
        "operationId": "CompanyController_addInvestment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InvestmentDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/getMinistries": {
      "get": {
        "operationId": "CompanyController_getMinistryUser",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/organisation/public/get": {
      "post": {
        "operationId": "CompanyController_getOrganisationPublicDetails",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Organisation"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": {
    "title": "National API",
    "description": "RESTful Web API Documentation.",
    "version": "0.5",
    "contact": {}
  },
...
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "api_key": { "type": "apiKey", "in": "header", "name": "api_key" }
    },
    "schemas": {
...
      "QueryDto": {
        "type": "object",
        "properties": {
          "page": { "type": "number" },
          "size": { "type": "number" },
          "filterAnd": {
            "type": "array",
            "example": [{ "key": "age", "operation": "gt", "value": 25 }],
            "items": { "$ref": "#/components/schemas/FilterEntry" }
          },
          "filterOr": {
            "type": "array",
            "example": [{ "key": "age", "operation": "gt", "value": 25 }],
            "items": { "$ref": "#/components/schemas/FilterEntry" }
          },
          "sort": {
            "type": "object",
            "example": { "key": "name", "order": "asc" },
            "items": { "$ref": "#/components/schemas/SortEntry" }
          },
          "filterBy": {
            "type": "object",
            "example": { "key": "ministryLevel", "values": ["1", "2"] },
            "items": { "$ref": "#/components/schemas/FilterBy" }
          },
          "extendedProperties": {
            "type": "object",
            "example": { "key": "isGetInvestmentHistory", "values": true }
          }
        },
        "required": ["page", "size"]
      },
      "DataExportQueryDto": {
        "type": "object",
        "properties": {
          "filterAnd": {
            "type": "array",
            "example": [{ "key": "age", "operation": "gt", "value": 25 }],
            "items": { "$ref": "#/components/schemas/FilterEntry" }
          },
          "filterOr": {
            "type": "array",
            "example": [{ "key": "age", "operation": "gt", "value": 25 }],
            "items": { "$ref": "#/components/schemas/FilterEntry" }
          },
          "sort": {
            "type": "object",
            "example": { "key": "name", "order": "asc" },
            "items": { "$ref": "#/components/schemas/SortEntry" }
          },
          "filterBy": {
            "type": "object",
            "example": { "key": "ministryLevel", "values": ["1", "2"] },
            "items": { "$ref": "#/components/schemas/FilterBy" }
          }
        }
      },
...
      "ByTypeDto": {
        "type": "object",
        "properties": { "companyRole": { "type": "object" } },
        "required": ["companyRole"]
      },
      "GetOrganizationsRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "IC",
              "PD",
              "API",
              "DNA",
              "Ministry",
              "ClimateFund",
              "ExecutiveCommittee"
            ]
          },
          "filterOwn": { "type": "boolean" }
        },
        "required": ["type", "filterOwn"]
      },
      "OrganisationSuspendDto": {
        "type": "object",
        "properties": { "remarks": { "type": "string" } }
      },
      "OrganisationRejectDto": {
        "type": "object",
        "properties": { "remarks": { "type": "string" } },
        "required": ["remarks"]
      },
      "FindOrganisationQueryDto": {
        "type": "object",
        "properties": {
          "companyIds": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["companyIds"]
      },
      "OrganisationUpdateDto": {
        "type": "object",
        "properties": {
          "companyId": { "type": "number" },
          "name": { "type": "string" },
          "taxId": { "type": "string" },
          "email": { "type": "string" },
          "paymentId": { "type": "string" },
          "website": { "type": "string" },
          "logo": { "type": "string" },
          "phoneNo": { "type": "string" },
          "faxNo": { "type": "string" },
          "address": { "type": "string" },
          "govDep": {
            "type": "string",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5",
              "6",
              "7",
              "8",
              "9",
              "10",
              "11",
              "12",
              "13",
              "14",
              "15",
              "16",
              "17",
              "18",
              "19",
              "20",
              "21",
              "22",
              "23",
              "24",
              "25",
              "26",
              "27",
              "28",
              "29",
              "30",
              "31",
              "32",
              "33",
              "34",
              "35",
              "36",
              "37",
              "38",
              "39",
              "40",
              "41",
              "42",
              "43",
              "44",
              "45",
              "46",
              "47",
              "48",
              "49",
              "50",
              "51",
              "52",
              "53",
              "54",
              "55",
              "56",
              "115",
              "57",
              "58",
              "59",
              "60",
              "61",
              "62",
              "63",
              "64",
              "65",
              "66",
              "67",
              "68",
              "69",
              "70",
              "71",
              "72",
              "73",
              "74",
              "76",
              "77",
              "78",
              "79",
              "80",
              "81",
              "82",
              "83",
              "84",
              "85",
              "86",
              "87",
              "88",
              "89",
              "90",
              "91",
              "92",
              "93",
              "94",
              "95",
              "96",
              "97",
              "98",
              "99",
              "100",
              "101",
              "102",
              "103",
              "104",
              "105",
              "106",
              "108",
              "109",
              "110",
              "111",
              "112",
              "113",
              "114",
              "116",
              "117",
              "118",
              "119",
              "120",
              "121",
              "122",
              "123",
              "124",
              "125",
              "126",
              "127",
              "128",
              "129",
              "130",
              "131",
              "132",
              "133",
              "134",
              "135",
              "136",
              "137",
              "138",
              "139",
              "140",
              "141",
              "142",
              "143",
              "144",
              "145",
              "146",
              "147",
              "148",
              "149",
              "150",
              "151",
              "152",
              "153",
              "154",
              "155",
              "156",
              "157",
              "158",
              "159"
            ]
          },
          "ministry": {
            "type": "string",
            "enum": [
              "Agriculture",
              "Aviation",
              "Communications",
              "Economy",
              "Education",
              "Energy",
              "Environment",
              "Health",
              "Intelligence",
              "Judiciary",
              "Maritime",
              "Media",
              "Science And Technology",
              "Water Resources",
              "Other"
            ]
          },
          "companyRole": {
            "type": "string",
            "enum": [
              "IC",
              "PD",
              "API",
              "DNA",
              "Ministry",
              "ClimateFund",
              "ExecutiveCommittee"
            ]
          },
          "regions": { "type": "array", "items": { "type": "string" } },
          "provinces": { "type": "array", "items": { "type": "string" } }
        },
        "required": [
          "companyId",
          "name",
          "taxId",
          "email",
          "paymentId",
          "address",
          "govDep",
          "ministry",
          "companyRole"
        ]
      },
      "InvestmentDto": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "nationalInvestmentId": { "type": "number" },
          "instrument": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Loan",
                "ResultBased",
                "Grant",
                "Equity",
                "Guarantee",
                "In-Kind",
                "Other",
                "ConcessionalLoan",
                "Non-ConcessionalLoan",
                "Insurance"
              ]
            }
          },
          "interestRate": { "type": "number" },
          "resultMetric": { "type": "string" },
          "paymentPerMetric": { "type": "number" },
          "startOfPayback": { "type": "number" },
          "guaranteePayback": {
            "type": "string",
            "enum": ["Upfront", "Annual"]
          },
          "insurancePayback": {
            "type": "string",
            "enum": ["LumpSum", "Annual"]
          },
          "period": { "type": "array", "items": { "type": "array" } },
          "comments": { "type": "string" },
          "type": { "type": "string", "enum": ["Public", "Private"] },
          "level": { "type": "string", "enum": ["National", "International"] },
          "stream": {
            "type": "string",
            "enum": ["ClimateFinance", "CarbonMarket"]
          },
          "esgClassification": {
            "type": "string",
            "enum": [
              "AAA",
              "AA",
              "A",
              "A(R)",
              "BBB",
              "BB",
              "B",
              "B(R)",
              "B13",
              "B13(R)",
              "CCC",
              "C",
              "C(R)",
              "FI"
            ]
          },
          "toCompanyId": { "type": "number" }
        },
        "required": ["amount", "nationalInvestmentId", "toCompanyId"]
      },
...
    }
  }
}